refactor: split settings manifest fields
